+++ client/package.json
@@ -13,7 +13,8 @@
     "@mantine/core": "^7.17.4",
     "@mantine/hooks": "^7.17.4",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "react-router-dom": "^7.5.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.21.0",
