+++ client/src/app/AppHeader.tsx
@@ -1,0 +1,7 @@
+export const AppHeader = () => {
+    return (
+        <div>
+            AppHeader
+        </div>
+    )
+}
