+++ client/src/pages/HomePage.tsx
@@ -1,0 +1,7 @@
+import {Button} from "@mantine/core";
+
+export const HomePage = () => {
+    return(
+        <Button variant="filled">Button</Button>
+    )
+}
